BAU-399 change formikSubmitHandler into useFormSubmit hook
