+++ src/explore-education-statistics-admin/src/pages/release/create-release/CreateReleasePage.tsx
@@ -2,6 +2,7 @@ import Link from '@admin/components/Link';
 import Page from '@admin/components/Page';
 import { ErrorControlState } from '@admin/contexts/ErrorControlContext';
 import withErrorControl from '@admin/hocs/withErrorControl';
+import useFormSubmit from '@admin/hooks/useFormSubmit';
 import ReleaseSummaryForm, {
   EditFormValues,
 } from '@admin/pages/release/summary/ReleaseSummaryForm';
@@ -14,7 +15,6 @@ import {
 } from '@admin/services/common/types';
 import service from '@admin/services/release/create-release/service';
 import { CreateReleaseRequest } from '@admin/services/release/create-release/types';
-import submitWithFormikValidation from '@admin/validation/formikSubmitHandler';
 import FormFieldRadioGroup from '@common/components/form/FormFieldRadioGroup';
 import {
   errorCodeAndFieldNameToFieldError,
@@ -29,6 +29,20 @@ import {
 import React, { useEffect, useState } from 'react';
 import { RouteComponentProps, withRouter } from 'react-router';
 import { ObjectSchemaDefinition } from 'yup';
+
+const errorCodeMappings = [
+  errorCodeToFieldError(
+    'SLUG_NOT_UNIQUE',
+    'timePeriodCoverageStartYear',
+    'Choose a unique combination of time period and start year',
+  ),
+  errorCodeAndFieldNameToFieldError(
+    'PARTIAL_DATE_NOT_VALID',
+    'NextReleaseDate',
+    'nextReleaseDate',
+    'Enter a valid date',
+  ),
+];
 
 interface MatchProps {
   publicationId: string;
@@ -66,33 +80,15 @@ const CreateReleasePage = ({
       .catch(handleApiErrors);
   }, [publicationId, handleApiErrors]);
 
-  const errorCodeMappings = [
-    errorCodeToFieldError(
-      'SLUG_NOT_UNIQUE',
-      'timePeriodCoverageStartYear',
-      'Choose a unique combination of time period and start year',
-    ),
-    errorCodeAndFieldNameToFieldError(
-      'PARTIAL_DATE_NOT_VALID',
-      'NextReleaseDate',
-      'nextReleaseDate',
-      'Enter a valid date',
-    ),
-  ];
+  const handleSubmit = useFormSubmit<FormValues>(async values => {
+    const createReleaseDetails: CreateReleaseRequest = assembleCreateReleaseRequestFromForm(
+      publicationId,
+      values,
+    );
 
-  const submitFormHandler = submitWithFormikValidation<FormValues>(
-    async values => {
-      const createReleaseDetails: CreateReleaseRequest = assembleCreateReleaseRequestFromForm(
-        publicationId,
-        values,
-      );
-
-      const createdRelease = await service.createRelease(createReleaseDetails);
-      history.push(summaryRoute.generateLink(publicationId, createdRelease.id));
-    },
-    handleApiErrors,
-    ...errorCodeMappings,
-  );
+    const createdRelease = await service.createRelease(createReleaseDetails);
+    history.push(summaryRoute.generateLink(publicationId, createdRelease.id));
+  }, errorCodeMappings);
 
   const cancelHandler = () =>
     history.push(appRouteList.adminDashboard.path as string);
@@ -154,7 +150,7 @@ const CreateReleasePage = ({
               ? Yup.string().required('Choose a template')
               : Yup.string(),
         })}
-        onSubmitHandler={submitFormHandler}
+        onSubmitHandler={handleSubmit}
         onCancelHandler={cancelHandler}
         additionalFields={
           model &&
