+++ src/explore-education-statistics-admin/src/pages/release/edit-release/ReleaseStatusPage.tsx
@@ -2,12 +2,12 @@ import ReleaseServiceStatus from '@admin/components/ReleaseServiceStatus';
 import StatusBlock from '@admin/components/StatusBlock';
 import { ErrorControlState } from '@admin/contexts/ErrorControlContext';
 import withErrorControl from '@admin/hocs/withErrorControl';
+import useFormSubmit from '@admin/hooks/useFormSubmit';
 import ManageReleaseContext, {
   ManageRelease,
 } from '@admin/pages/release/ManageReleaseContext';
 import permissionService from '@admin/services/permissions/service';
 import service from '@admin/services/release/edit-release/status/service';
-import submitWithFormikValidation from '@admin/validation/formikSubmitHandler';
 import Button from '@common/components/Button';
 import ButtonText from '@common/components/ButtonText';
 import { Form, FormFieldRadioGroup, Formik } from '@common/components/form';
@@ -19,6 +19,14 @@ import { ReleaseStatus } from '@common/services/publicationService';
 import { FormikProps } from 'formik';
 import React, { useContext, useEffect, useState } from 'react';
 import { RouteComponentProps } from 'react-router';
+
+const errorCodeMappings = [
+  errorCodeToFieldError(
+    'APPROVED_RELEASE_MUST_HAVE_PUBLISH_SCHEDULED_DATE',
+    'releaseStatus',
+    'Enter a publish scheduled date before approving',
+  ),
+];
 
 interface FormValues {
   releaseStatus: ReleaseStatus;
@@ -82,33 +90,23 @@ const ReleaseStatusPage = ({
       .catch(handleApiErrors);
   }, [releaseId, handleApiErrors, showForm]);
 
-  if (!model) return null;
-
-  const formId = 'releaseStatusForm';
-
-  const errorCodeMappings = [
-    errorCodeToFieldError(
-      'APPROVED_RELEASE_MUST_HAVE_PUBLISH_SCHEDULED_DATE',
-      'releaseStatus',
-      'Enter a publish scheduled date before approving',
-    ),
-  ];
-
-  const submitFormHandler = submitWithFormikValidation<FormValues>(
-    async values => {
-      await service.updateReleaseStatus(releaseId, values).then(() => {
+  const handleSubmit = useFormSubmit<FormValues>(async values => {
+    await service.updateReleaseStatus(releaseId, values).then(() => {
+      if (model) {
         setModel({
           ...model,
           releaseStatus: values.releaseStatus,
           statusOptions: model.statusOptions,
         });
+      }
 
-        setShowForm(false);
-      });
-    },
-    handleApiErrors,
-    ...errorCodeMappings,
-  );
+      setShowForm(false);
+    });
+  }, errorCodeMappings);
+
+  if (!model) return null;
+
+  const formId = 'releaseStatusForm';
 
   return (
     <>
@@ -142,7 +140,7 @@ const ReleaseStatusPage = ({
             releaseStatus: model.releaseStatus,
             internalReleaseNote: '',
           }}
-          onSubmit={submitFormHandler}
+          onSubmit={handleSubmit}
           validationSchema={Yup.object<FormValues>({
             releaseStatus: Yup.mixed().required('Choose a status'),
             internalReleaseNote: Yup.string().required(
